added tests for XMLMemento (reusing existing JSONMemento tests)
