+++ framework/tests/org.jboss.tools.rsp.launching.test/resources/mementoTest.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="UTF-8" standalone="no"?>
+<root rootPropKey="rootPropValue">
+   <childOne childStringKey="childStringValue" />
+   <childTwo childStringKey="childStringValue" />
+</root>
